+++ src/fields/BaseRelationField.php
@@ -534,6 +534,7 @@ abstract class BaseRelationField extends Field implements PreviewableFieldInterf
     public function modifyElementIndexQuery(ElementQueryInterface $query): void
     {
         $criteria = [
+            'drafts' => null,
             'status' => null,
         ];
 
@@ -910,9 +911,7 @@ JS;
     protected function inputTemplateVariables($value = null, ?ElementInterface $element = null): array
     {
         if ($value instanceof ElementQueryInterface) {
-            $value = $value
-                ->status(null)
-                ->all();
+            $value = $value->all();
         } else if (!is_array($value)) {
             $value = [];
         }
@@ -1088,6 +1087,7 @@ JS;
     {
         $clone = clone $query;
         $clone
+            ->drafts(null)
             ->status(null)
             ->siteId('*')
             ->limit(null)